Extract InboxScoped concern so tab controllers stop inheriting from InboxesController
