+++ app/controllers/inboxes/clearances_controller.rb
@@ -1,4 +1,4 @@
-class Inboxes::ClearancesController < InboxesController
+class Inboxes::ClearancesController < ApplicationController
   def create
     if params[:scope] == "direct_messages"
       Current.user.mark_direct_messages_as_read(session[:inbox_last_loaded_dms_created_at])
