+++ app/controllers/inboxes/messages_controller.rb
@@ -1,4 +1,6 @@
-class Inboxes::MessagesController < InboxesController
+class Inboxes::MessagesController < ApplicationController
+  include InboxScoped
+
   before_action :set_message_pagination_anchors, if: :paginating?
 
   def index
